add login_required decorator

+++ base/views.py
@@ -1,5 +1,7 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
@@ -33,12 +35,14 @@ def loginPage(request):
         else:
             messages.error(request, 'Username OR password is incorrect')
 
-    context ={}
+    context = {}
     return render(request, 'base/login_register.html', context)
+
 
 def logoutUser(request):
     logout(request)
     return redirect('home')
+
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') is not None else ''
@@ -46,7 +50,7 @@ def home(request):
         Q(topic__name__icontains=q) |
         Q(name__icontains=q) |
         Q(description__icontains=q)
-        )
+    )
 
     topics = Topic.objects.all()
     room_count = rooms.count()
@@ -54,11 +58,14 @@ def home(request):
     context = {'rooms': rooms, 'topics': topics, 'room_count': room_count}
     return render(request, 'base/home.html', context)
 
+
 def room(request, pk):
     room = Room.objects.get(id=pk)
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
+
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -70,9 +77,14 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed to do that')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -83,9 +95,16 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed to do that')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
+
     return render(request, 'base/delete.html', {'obj': room})
